+++ lib/mcumgr_flutter.dart
@@ -7,6 +7,8 @@ import 'models/image_upload_alignment.dart';
 import 'models/firmware_upgrade_mode.dart';
 import 'src/mcumgr_update_manager.dart';
 
+export 'src/mcumgr_settings.dart';
+
 part 'src/mcumgr_flutter.dart';
 part 'models/progress_update.dart';
 part 'models/mcu_log_message.dart';